service-graphql-user/requirements.txt
Flask==3.0.0
strawberry-graphql[flask]==0.216.0